MoviesList & PopularMovies adjusted
